+++ src/utils/vueTsr.ts
@@ -1,17 +1,16 @@
 import fs from 'fs'
 import { createSSRApp } from 'vue'
 import { renderToString } from '@vue/server-renderer'
+import { uriFor } from '../../templates/utils/uri'
 
-export function vueToHtml(
-    tmpStr: string,
-    data: any,
-    methods?: any
-): Promise<string> {
+export function vueToHtml(tmpStr: string, data: any): Promise<string> {
     const m = tmpStr.match(/<template>([\s\S]*?)<\/template>/)
     const app = createSSRApp({
         template: m![1],
         data: () => data,
-        methods,
+        methods: {
+            uriFor,
+        },
     })
     return renderToString(app)
 }