+++ webpack.config.js
@@ -1,7 +1,18 @@
 const { merge } = require('webpack-merge')
 const common = require('./webpack.common.js')
+const WebpackObfuscator = require('webpack-obfuscator')
 
 module.exports = merge(common, {
-    mode: 'production'
+    mode: 'production',
     // mode: 'development',
+    plugins: [
+        new WebpackObfuscator(
+            {
+                trasnformObjectKeys: true,
+                rotateStringArray: true,
+                roatetStringArrayEnable: true,
+            },
+            ['excluded_bundle_name.js']
+        ),
+    ]
 });